# server/src/swagger.json
{
    "swagger": "2.0",
    "info": {
      "version": "2.0",
      "title": "Movie Booking API",
      "description": "Document API",
      "license": {
        "name": "Movie Booking",
        "url": "https://opensource.org/licenses/MIT"
      }
    },
    "host": "localhost:1000",
    "basePath": "/",
...
    "schemes": [
      "http",
      "https"
    ],
    "consumes": [
      "application/json"
    ],
    "produces": [
      "application/json"
    ],
    "securityDefinitions": {
        "ApiKeyAuth":{
          "type": "apiKey",
          "in": "headers",
          "name": "authorization"
        }
    },
    "paths": {
...
      "/api/news": {
        "get": {
          "summary": "Get all news",
          "tags": [
            "User"
          ],
          "description": "Get all news",
          "parameters": [
          ],
          "produces": [
            "application/json"
          ],
          "responses": {
            "200": {
              "description": "Get all news Success!",
              "schema": {
                "$ref": "#/definitions/News"
              }
            },
            "404":{
              "description": "Error Get all news!"
            }
          }
        },
        "post": {
          "summary": "Create new news",
          "tags": [
            "User"
          ],
          "description": "Create new news",
          "parameters": [
            {
              "name": "body",
              "in": "body",
              "description": "Create new news",
              "schema": {
                "$ref": "#/definitions/News"
              }
            }
          ],
          "produces": [
            "application/json"
          ],
          "responses": {
            "200": {
              "description": "Create new news Success!",
              "schema": {
                "$ref": "#/definitions/News"
              }
            },
            "404":{
              "description": "Error Create new news!"
            }
          }
        }
      },
      "/api/news/:id": {
        "get": {
          "summary": "Get news by id",
          "tags": [
            "User"
          ],
          "description": "Get news by id",
          "parameters": [
            {
              "name": "Id",
              "in": "path",
              "description": "Get news by id",
              "type": "string"
            }
          ],
          "produces": [
            "application/json"
          ],
          "responses": {
            "200": {
              "description": "Get news by id Success!",
              "schema": {
                "$ref": "#/definitions/News"
              }
            },
            "404":{
              "description": "Error Get news by id!"
            }
          }
        },
        "put": {
          "summary": "Update news by id",
          "tags": [
            "User"
          ],
          "description": "Update news by id",
          "parameters": [
            {
              "name": "body",
              "in": "body",
              "description": "Update news by id",
              "schema": {
                "$ref": "#/definitions/News"
              }
            }
          ],
          "produces": [
            "application/json"
          ],
          "responses": {
            "200": {
              "description": "Update news by id Success!",
              "schema": {
                "$ref": "#/definitions/News"
              }
            },
            "404":{
              "description": "Error Update news by id film!"
            }
          }
        },
        "delete": {
          "summary": "Delete news by id",
          "tags": [
            "User"
          ],
          "description": "Delete news by id",
          "parameters": [
            {
              "name": "Id",
              "in": "path",
              "description": "Delete news by id",
              "type": "string"
            }
          ],
          "produces": [
            "application/json"
          ],
          "responses": {
            "200": {
              "description": "Delete news by id Success!"              
            },
            "404":{
              "description": "Error Delete news by id!"
            }
          }
        }
      },
...
      "/api/ticket/ticket/:id": {
        "get": {
          "summary": "Get all seat_ids",
          "tags": [
            "User"
          ],
          "description": "Get all seat_ids",
          "parameters": [
          ],
          "produces": [
            "application/json"
          ],
          "responses": {
            "200": {
              "description": "Get all seat_ids Success!",
              "schema": {
                "type":"array",
                "items":["string","string"],
                "example":["A10","B9"]
              }
            },
            "404":{
              "description": "Error Get all seat_ids!"
            }
          }
        }
      },
      "/api/ticket": {
        "get": {
          "summary": "Get all ticket",
          "tags": [
            "User"
          ],
          "description": "Get all ticket",
          "parameters": [
          ],
          "produces": [
            "application/json"
          ],
          "responses": {
            "200": {
              "description": "Get all ticket Success!",
              "schema": {
                "$ref": "#/definitions/Ticket"
              }
            },
            "404":{
              "description": "Error Get all ticket!"
            }
          }
        },
        "post": {
          "summary": "Create new ticket",
          "tags": [
            "User"
          ],
          "description": "Create new ticket",
          "parameters": [
            {
              "name": "body",
              "in": "body",
              "description": "Create new ticket",
              "schema": {
                "$ref": "#/definitions/Ticket"
              }
            }
          ],
          "produces": [
            "application/json"
          ],
          "responses": {
            "200": {
              "description": "Create new ticket Success!",
              "schema": {
                "$ref": "#/definitions/Ticket"
              }
            },
            "404":{
              "description": "Error Create new ticket!"
            }
          }
        }
      },
      "/api/ticket/:id": {
        "get": {
          "summary": "Get ticket by id",
          "tags": [
            "User"
          ],
          "description": "Get ticket by id",
          "parameters": [
            {
              "name": "Id",
              "in": "path",
              "description": "Get ticket by id",
              "type": "string"
            }
          ],
          "produces": [
            "application/json"
          ],
          "responses": {
            "200": {
              "description": "Get ticket by id Success!",
              "schema": {
                "$ref": "#/definitions/Ticket"
              }
            },
            "404":{
              "description": "Error Get ticket by id!"
            }
          }
        },
        "put": {
          "summary": "Update ticket by id",
          "tags": [
            "User"
          ],
          "description": "Update ticket by id",
          "parameters": [
            {
              "name": "body",
              "in": "body",
              "description": "Update ticket by id",
              "schema": {
                "$ref": "#/definitions/Ticket"
              }
            }
          ],
          "produces": [
            "application/json"
          ],
          "responses": {
            "200": {
              "description": "Update ticket by id Success!",
              "schema": {
                "$ref": "#/definitions/Ticket"
              }
            },
            "404":{
              "description": "Error Update ticket by id film!"
            }
          }
        },
        "delete": {
          "summary": "Delete ticket by id",
          "tags": [
            "User"
          ],
          "description": "Delete ticket by id",
          "parameters": [
            {
              "name": "Id",
              "in": "path",
              "description": "Delete ticket by id",
              "type": "string"
            }
          ],
          "produces": [
            "application/json"
          ],
          "responses": {
            "200": {
              "description": "Delete ticket by id Success!"              
            },
            "404":{
              "description": "Error Delete ticket by id!"
            }
          }
        }
      },
...
      "/api/theater/detail/:id": {
        "get": {
          "summary": "Get all theaters by id",
          "tags": [
            "User"
          ],
          "description": "Get all theaters by id",
          "parameters": [
          ],
          "produces": [
            "application/json"
          ],
          "responses": {
            "200": {
              "description": "Get all theaters by id Success!",
              "schema": {
                "$ref": "#/definitions/Theater"
              }
            },
            "404":{
              "description": "Error Get all theaters by id!"
            }
          }
        }
      },
      "/api/theater/getfilmtoday": {
        "get": {
          "summary": "Get all film today from theater",
          "tags": [
            "User"
          ],
          "description": "Get all film today from theater",
          "parameters": [
          ],
          "produces": [
            "application/json"
          ],
          "responses": {
            "200": {
              "description": "Get all film today from theater Success!",
              "schema": {
                "type":"array",
                "items":{
                  "type":"object",
                  "properties":{
                    "_id": {"type":"string","format":"ObjectId"},
                    "name": {"type":"string"},
                    "address": {"type":"string"},
                    "url_image": {"type":"string"},
                    "films": {
                        "type":"object",
                        "properties":{
                          "_id": "5ff147f263ccaf2a14e45106",
                        "name":{"type":"string"},
                        "trailer": {"type":"string"},
                        "countries": {"type":"string"},
                        "long_time": {"type":"string"},
                        "start_date":{"type":"string", "format":"date"},
                        "rates": {"type":"string"},
                        "rate_count": {"type":"string"},
                        "url_avatar":{"type":"string"},
                        "url_background": {"type":"string"},
                        "is_blockbuster":{"type":"string"},
                        "film_schedules": 
                            {
                            "type":"object",
                            "properties":{
                            "_id": {"type":"string","format":"ObjectId"},
                            "time_start": {"type":"string", "format":"date"},
                            "time_end": {"type":"string", "format":"date"},
                            "film_id": {"type":"string","format":"ObjectId"},
                            "theater_id": {"type":"string","format":"ObjectId"},
                            "room":{"type":"string"}
                            }
                            }
                          } 
                        }
                        
                  }
                }
              }
            },
            "404":{
              "description": "Error Get all film today from theater!"
            }
          }
        }
      },
      "/api/theater/:id/detail": {
        "get": {
          "summary": "Get all film 7 day by id",
          "tags": [
            "User"
          ],
          "description": "Get all film 7 day by id",
          "parameters": [
          ],
          "produces": [
            "application/json"
          ],
          "responses": {
            "200": {
              "description": "Get all film 7 day by id Success!",
              "schema": {
                "type":"array",
                "items":{
                  "type":"object",
                  "properties":{
                    "_id": {"type":"string","format":"ObjectId"},
                    "name": {"type":"string"},
                    "address": {"type":"string"},
                    "url_image": {"type":"string"},
                    "films": {
                        "type":"object",
                        "properties":{
                          "_id": "5ff147f263ccaf2a14e45106",
                        "name":{"type":"string"},
                        "trailer": {"type":"string"},
                        "countries": {"type":"string"},
                        "long_time": {"type":"string"},
                        "start_date":{"type":"string", "format":"date"},
                        "rates": {"type":"string"},
                        "rate_count": {"type":"string"},
                        "url_avatar":{"type":"string"},
                        "url_background": {"type":"string"},
                        "is_blockbuster":{"type":"string"},
                        "film_schedules": 
                            {
                            "type":"object",
                            "properties":{
                            "_id": {"type":"string","format":"ObjectId"},
                            "time_start": {"type":"string", "format":"date"},
                            "time_end": {"type":"string", "format":"date"},
                            "film_id": {"type":"string","format":"ObjectId"},
                            "theater_id": {"type":"string","format":"ObjectId"},
                            "room":{"type":"string"}
                            }
                            }
                          } 
                        }
                        
                  }
                }
              }
            },
            "404":{
              "description": "Error Get all film 7 day by id!"
            }
          }
        }
      },
      "/api/theater": {
        "get": {
          "summary": "Get all theaters",
          "tags": [
            "User"
          ],
          "description": "Get all theaters",
          "parameters": [
          ],
          "produces": [
            "application/json"
          ],
          "responses": {
            "200": {
              "description": "Get all theaters Success!",
              "schema": {
                "$ref": "#/definitions/Theater"
              }
            },
            "404":{
              "description": "Error Get all theaters!"
            }
          }
        },
        "post": {
          "summary": "Create new theaters",
          "tags": [
            "User"
          ],
          "description": "Create new theaters",
          "parameters": [
            {
              "name": "body",
              "in": "body",
              "description": "Create new theaters",
              "schema": {
                "$ref": "#/definitions/Theater"
              }
            }
          ],
          "produces": [
            "application/json"
          ],
          "responses": {
            "200": {
              "description": "Create new theaters Success!",
              "schema": {
                "$ref": "#/definitions/Theater"
              }
            },
            "404":{
              "description": "Error Create new theaters!"
            }
          }
        }
      },
      "/api/theaters/:id": {
        "get": {
          "summary": "Get theaters by id",
          "tags": [
            "User"
          ],
          "description": "Get theaters by id",
          "parameters": [
            {
              "name": "Id",
              "in": "path",
              "description": "Get theaters by id",
              "type": "string"
            }
          ],
          "produces": [
            "application/json"
          ],
          "responses": {
            "200": {
              "description": "Get theaters by id Success!",
              "schema": {
                "$ref": "#/definitions/Theater"
              }
            },
            "404":{
              "description": "Error Get theaters by id!"
            }
          }
        },
        "put": {
          "summary": "Update theaters by id",
          "tags": [
            "User"
          ],
          "description": "Update theaters by id",
          "parameters": [
            {
              "name": "body",
              "in": "body",
              "description": "Update theaters by id",
              "schema": {
                "$ref": "#/definitions/Theater"
              }
            }
          ],
          "produces": [
            "application/json"
          ],
          "responses": {
            "200": {
              "description": "Update theaters by id Success!",
              "schema": {
                "$ref": "#/definitions/Theater"
              }
            },
            "404":{
              "description": "Error Update theaters by id film!"
            }
          }
        },
        "delete": {
          "summary": "Delete theaters by id",
          "tags": [
            "User"
          ],
          "description": "Delete theaters by id",
          "parameters": [
            {
              "name": "Id",
              "in": "path",
              "description": "Delete theaters by id",
              "type": "string"
            }
          ],
          "produces": [
            "application/json"
          ],
          "responses": {
            "200": {
              "description": "Delete theaters by id Success!"              
            },
            "404":{
              "description": "Error Delete theaters by id!"
            }
          }
        }
      },
...
    },
    "definitions": {
...
      "News":{
        "properties":{
          "title":{"type":"string"},
          "content":{"type":"string"},
          "hastag":{"type":"Array"},
          "public_date":{"type":"string"," format":"date"},
          "film_id":{"type":"string"},
          "is_deleted":{"type":"boolean"},
          "created_at":{"type":"string"," format":"date"},
          "updated_at":{"type":"string"," format":"date"}
        }
      },
...
      "Theater":{
        "properties":{
          "name":{"type":"string"},
          "address":{"type":"string"},
          "url_image":{"type":"string"},
          "comment":{"type":"string"},
          "is_deleted":{"type":"boolean"},
          "created_at":{"type":"string"," format":"date"},
          "updated_at":{"type":"string"," format":"date"}
        }
      },
...
      "Ticket":{
        "properties":{
          "customer_id":{"type":"string"},
          "film_id":{"type":"string"},
          "count":{"type":"Number"},
          "voucher_id":{"type":"string"},
          "seat":{"type":"Array"},
          "booking_time":{"type":"string"," format":"date"},
          "cost":{"type":"Number"},
          "is_deleted":{"type":"boolean"},
          "created_at":{"type":"string"," format":"date"},
          "updated_at":{"type":"string"," format":"date"}
        }
      },
...
    }
  }